Data Byte Flattening
Mentioned this in the previous commit, making a type-safe, less complex implementation for the reading/writing backings. I wanted to alleviate some of the declaration duplication I had created with the original setup I had. While it had less abstraction for how the methods themselves work, it kind of made things less straightforward in terms of trying to see the pattern between how they are all implemented. What do you think, is this more readible? Or should I use the less abstract implementation? This is more simplified, but it's also less modular, in terms of specifying certain behavior for each method on it's own. It's like the same trade-off from the video I mentioned in the previous commit.

https://www.youtube.com/watch?v=CXyzmS2eNFo