SNBT Parse Temp Migration
I started trying to port the SNBT parser over to a functional API as well, but it started to get too complicated, and like other times I've tried to update it, I'm not sure how to upgrade things of it, just yet. I'm happy that I have it as part of my test setup though, because I don't think I've actually tried adjusting parts of it's implementation for fear of worrying I'd break something about it's parsing abilities. I think I'm going to try that concept with it instead, rather than implement parts of it piece by piece.

https://www.reddit.com/r/functionalprogramming/comments/zt2wnr/functional_implementation_of_a_parser/
https://softwareengineering.stackexchange.com/questions/213345/how-to-refactor-an-oo-program-into-a-functional-one
https://news.ycombinator.com/item?id=21995176
https://stackoverflow.com/questions/4564414/delete-first-character-of-string-if-it-is-0
https://www.measurethat.net/Benchmarks/Show/2335/1/slice-vs-substr-vs-substring-with-no-end-index (ended up going with `.slice()` because I'm familar with it, and it's the same one I use for TypedArray objects too, so using the same one across all of them helps with understanding what the concept is, for consistency.)

And yeah, this code here is currently just a copy of the existing parse code, without any edits made yet.