SNBT Parser: I Field Removal
Was looking into this library to see if it might have any other good things to add/look into for my SNBT parser, but it doesn't seem to be quite what I'm looking for. Still a nice find! I think this version is a bit closer to the setup that I want to have, and that one is still in the class setup I'm trying to migrate from.

https://www.npmjs.com/package/@bdsx/snbt
https://github.com/bdsx/bdsx

Invisible Face! KG&TLW is a great band, I can't believe how much they've released over the years, it is inspiring for me to keep trying to do that as well.