SNBT Parser: Class Removal!
Yaay!!! Been wanting to have this for some time now. Slow but surely chipping away at making my own SNBT parser from scratch. It's the challenging part of flattening the tree of how things are added on top of each other, like in terms of `this` usage, or state manipulation. Now it is cleanly described for each things in terms of being on a functional level. I'm glad the NBT reader and writer backings allow for that as well.

Classless type concept thing, I thought it was interesting that you can have both a type and a function of the same name. I guess that's no different than the `as const`/`type` setup though.

```ts
interface NBTData {
  fart: number;
}

const data = NBTData();

function NBTData({ fart = 5 }: Partial<NBTData> = {}): NBTData {
  return { fart };
}
```