FP-Reader-Writer Merge
Merging everything from `index-new.ts` into the regular source for the whole project. Going to re-add the missing validation features in the next commit, to show the difference between what the rework file was missing, and what it still needs as a library to be secure again.

Normalizer 2 solo, on it's own 🎶 🥁