Original Docs + Param Validation (Write needs work)
Bringing back some of the missing regular features that didn't get carried over to the rewrite, because I didn't want to cloud my mind with the extra config things that are more for API management than actual program logic.

Discovered that there was an issue with the BedrockLevel header detection, which didn't account for when a given file is smaller than the BedrockLevel header, which would error out because the reader would try to read from bytes that didn't exist in that short of a buffer. I added a new test file to account for that, which seems to be the smallest possible file that NBT can be. It's an unnamed, empty Compound tag. Only two bytes, extra small!

I still need to properly test, and re-implement the validation for `write()`. It seems to need a better restructure because of the API concept changes. I want to ensure it's right.