Write Validation Returns!
Man, I was losing my brain a little bit. Some of the generic usage around the `object` type can get a bit confusing at times. This strengthens the usage of `NBTData`, so it is what is used to consistently package things throughout the API. For the more user-accessible parts of the API, that's going to allow options objects. But for more internal usage ones, I think I may use `NBTData` instead. This may all be up in the air though, we'll see. Not absolutely sure where and how I want to validate things safely, consistently, and without bloating the code a bunch all over the place.