Writer Builder Usage
Not completely sure this is how things completely are meant to work, but it's to help with building things simpler. Inspired by help in discussion from @peacefulotter.

https://github.com/peacefulotter/minecraft-server
https://wiki.vg/Protocol#Data_types

I may be trying to broaden the use of NBTify too far by exposing these methods as part of the API, I'm not sure how I feel about that yet.

I do think I need something similar for my MCRegionJS project though, since LCE uses a custom non-NBT binary format for most chunk versions.
https://github.com/Offroaders123/MCRegionJS

This seems to be the same concept that UtterEvergreen's and Jerrin's projects use as well, with the `DataInputManager` and `DataOutputManager` classes (maybe it's just `DataManager`).
https://github.com/zugebot/LegacyEditor

This is the general concept for it:

```ts
// the API is weird, I still need to adjust it to look right, since it's still internally-shaped atm.
NBT.writeDouble(undefined, 5).trimmedEnd();
```

I also still have to test if building plain buffers works like this. I still have to do that with the SNBT ones too, as those I definitely want to make exposable. Dovetail will use that for the UI, when editing tags from the tree view GUI. I want to add a dialog popup that shows you the editable SNBT just for that plain tag. That will make it really simple to derive the UI from that too, because it's just a simple parse step, rather than a class kind of thing. NBTify's use of primitives makes it kind of interesting to build the UI from it. I think using string indexing for keys and such is probably the way I'll go for it.