Async Iterator Fallback
Looked more into the browser support for this, and Caniuse seems to be a little behind on it. It says only very new versions of Chrome support it, and Firefox. I thought it's been around longer than that though? Either way, Node supports it too (probably because of the Chrome part, I don't think it supported it before either, so maybe it did just get added to Chrome? I'm confused lol), but Safari still doesn't.

I demoed out the `for await` loop here in all three browsers, and Node, and everything as of now supports it, except for Safari. The metrics for that only means the latest versions of the environments though, so I'm not going to call it a complete dive in. I'm still going to use the polyfill I have. TS doesn't have types for it yet either, so that's another one to look out for. Once TS has types for things, that usually means it's supported in a larger variety of places.

https://www.reddit.com/r/webdev/comments/11n4h7y/stream_is_not_async_iterable/
https://jakearchibald.com/2017/async-iterators-and-generators/
https://caniuse.com/mdn-api_readablestream_--asynciterator

```html
<!DOCTYPE html>
<script type="module">
  console.clear();
  console.log("hi");
  try {
    for await (const data of (await fetch("https://stedit.app")).body) { console.log(data); }
  } catch (error) {
    console.error(error);
  }
  console.log(ReadableStream.prototype[Symbol.asyncIterator]);
</script>
```