Stringify Primitive Flattening
Wanted to fix this a while back, glad it can be possible now! I don't have to construct the primitive wrappers to directly write a given value as a specific SNBT type. I'm not sure this did much in terms of performance (making unnecessary objects for each array item, to just be discarded over again), but it was definitely unnecessary. This is the kind of thing I think ThePrimeagen mentions to look out for, making things in unnecessary or repetitive times.

https://minecraft.wiki/w/NBT_format#SNBT_format