Prismarine-NBT Converter Demo
I used this concept in part of what I think was Region-Types, a few months ago. I was curious what this might look like as a concept if it were built into NBTify itself. The idea is that you'd be able to have a pre-existing Prismarine-NBT parsed NBT file object, and NBTify would know how to convert those shapes into the object structure that it uses, like it's primitives and things like that. Ideally I would just map the values themselves, rather than reencoding the buffers like this, I was curious just how close this would get though, since it has worked for me already, it just might be more extra computations that don't have to be there.

What do you think? Is writing it back to a file, and re-reading it better than just mapping over to the other object shapes? Maybe this really is just faster. I have no data to back up which one is better, but I do like the simplicity of this, it works well with the possibility of combining with the other libraries as well. This is a niche thing though, so I'm not sure I'm going to add this anyways.