CLI Typings
With removing the value validations from the `NBTData` constructor, I need to properly validate the values on my own elsewhere throughout the package. Going to work on strengthening the project as a whole, and try to put a new set of eyes on things again, as to how they should work.