Optional Spacing
I think this is kind of annoying, but more consistent at the same time. Both `JSON.stringify()` and `NBT.stringify()` don't format by default, so I question whether the CLI doing that would make any sense. From a UX perspective I think it makes it easier for the user, but should it be different than how the programming side of it works?