TAG Byte Reading Validation
Realized I shouldn't be making assertions at this part. This can make things more secure!

I want to look through the code again after this, and refactor it, but not to move things around, but to see if there's any weird things still lying around the codebase.

Solar Winds, it kind of sounds like Pineapple Thief at the beginning :D

Listened to Insurgentes and Grace for Drowning disc 1 today :D