Primitive Typing Strictness
Going to go back through the library and make it stronger in missing places, as well as fixing places that I changed along the way, unknowingly.

I didn't have the return types here because I thought the assertions could just be enough to declare what they were, but now I think I'm full on board with using type annotations for things, it definitely helps you keep track of your intent a lot better, which is what I like it for, to keep track of what is temporary, and what needs work. If something has a type annotation, it's more likely to be marked as something complete, in my head.