Complete Full Typings
I'm not sure what I can do specifically for this yet, this project feels like a complete disaster to me now at this point, kind of like how STE felt for a bit, before the rehash. Even that feels a bit of a mess even with the SolidJS migration though.

I want to make the code quality for this much better, but I'm not completely sure yet what specific areas should be changed, I know things just feel very off though. I think the scopes of my functions are too big, and the typings are too complex at the low level implementation. I want to make it easier to understand the underlying implementation of things, but still allow the overarching API to be modular for users.

My first thought was to set up the project with complete typings for existing implicit types, so I can help myself understand what needs the most work here. I think I either try to build onto things from the bottom up, or the top down, and it doesn't seem to work either way. Starting fresh leads to forgetful code, and building on top of existing things makes it a mess. Maybe I just need to refactor the scope of things? That's my current feeling. It's hard to adjust the code when you notice your making the same changes that let the project to this point in the first place. It's like everything now is exactly how I wanted it to be back then, but now that we're here, that's not cutting it for me. A very interesting situation, and I think it means/is pointing me towards that I have plenty more to learn in terms of structuring implementations and scoping of things.