NBTError Removal
Removing this in favor of looking into a better solution later down the road.

This is breaking for Bedrock-LevelDB, but I also don't want to rely on how this is currently setup, I intend to start making the code for this much better, and in doing that I have to make breaking changes where it can lead to success later.

#31