Big Updates 🚀
Ok, I think I'm going to move up to v2.0.0!

This is mostly because I plan to make API changes fairly soon, and a lot of progress has been made for this project over the last few months (relatively haha, I don't think the progress has been as speedy as when it first started though)

I'd say the biggest set of changes (going through the diffs, there's so many to count, aaah!) are:

- little-varint support! (added today)
- SNBT reading and writing are now functional rather than class-based (might go back on one of these, we'll see). The main thing for wanting to do this is to allow projects like Dovetail to have access to parsing tags of any kind, rather than only root tags, say for input elements for the NBT tree view.
- MUTF-8 support
- Tag type generics (helpful for Region-Types)
- ArrayTag aliases (say like passing a `Uint8Array` into NBTify, it will accept it as a `IntArrayTag` now! Not exclusively `Int8Array`)
- Removed `NBTError` in favor of figuring out a better solution for it later
- Improved CLI file type resolutions

I went back on a few of the changes from the last npm version (or two), having to do with the NBTData prop validations, and things around `StorageVersion` and `bedrockLevel`.

I this this publish took so long because I had started experimenting with all of the implementation backing refactorings, and I ended up reverting a bunch of them, because they seemed better in concept than in execution. Then I kind of lost steam for working on the project, because that left me stranded not knowing where I was at, and whether I was moving forward with it. With a little work later on though, I ended up diffing the remnants out again, and now we are where we have things now. That let me start brainstorming new and existing issue fixes again.

Been going back and listening to my old music from over the last few years, listening to "Receptacle / Incinerator / Demolisher / Coffee Cake" right now. Lots cooler stuff than I remembered ever writing! Lost of lore lol.