Adjacent NBT Reading
Ziltoid the Omniscient!!!

#31
#51

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/fromAsync

This is a demo of this feature, I'm not sure if this is completely how I will implement it, but it does seem to be working fairly okay.

The `chunk91` file is from Dexrn, in the project to parse CDB entries from New 3DS Edition. It seems to be relatively the same format as that of the equivalent format section in Bedrock's LevelDB implementation. That's where my `BlockEntity` file is from too, by Bedrock-LevelDB project.