Strictless Output Persistence
#31
#51

https://lifesigns1.bandcamp.com/album/lifesigns

Funny how in the first issue, I wasn't liking the new GitHub UI at the time, now I don't mind it, and it does look really nice :D

This first one is specifically for the persistence of the read offset when `strict` mode is disabled. It is only accessible with a getter on the NBTData object, and it will return `null` otherwise, when strict mode is enabled (still on by default).

This is initially to set up allowing for easier test-parsing of files with adjacent content, like Bedrock's LevelDB entries. With NBTify currently not easily making the previous offset accessible, it makes it hard to deduce where the next reading step should start from. This simply adds that transparency on it's own.

I may rework the method by how the NBTData object handles this, make it a little tidier.

https://github.com/Offroaders123/NBTify/commit/4664e4c1bee54c99819f768c3996ed62408fe59f