Branded Primitive Classes
This seems to be a much more useable API in terms of the TypeScript side of things. It's completely symmetrical in terms of the JavaScript implementation, so nothing changes there. It essentially just allows you to type-safely use a `Int8` for example, just as a regular `number` type. This may sound dangerous in some respects, and yes this is just a concept, so I plan to look into it to see if anything is wrong with doing it like this. Essentially, JS already converts `number` primitives back and forth for you, say like when you do `(5).toString()` for example. It's making a primitive wrapper object of that value, for a single line, and getting the result of that method call, from that object that was created under the hood.

So the concept here is to use branded types to distinguish apart the number primitives, and use real `Number`-extended wrapper objects in runtime to distinguish them apart there as well. Being able to extend `Number` also allows you to pass them around and use them safely as plain `number` values too, since JS is able to do this conversion for us under the hood already. Like if you do `new Number(5) + 2`, you get `7`. TypeScript will warn you of doing this, and that has been my issue with using the wrappers for regular scripting purposes. I wanted to be able to pass them around as pure values, and not have to call `valueOf()` every single time I access their values.

Another side to this though, is that this does notable break the concerns that TS provides, even though this is very much nearly doable in most JS cases. Say like you use `typeof myByte === "number"`, it will be `false` because it's actually an `object`. So maybe it makes sense to explicitly unwrap the value when you completely need to, and if not, you can still safely narrow the type and use addition for example, and decide on your own when it is safe to unwrap the type or not. I can provide more examples for this later on. I think this may be the way I take things, I really like that it allows for the use of regular function calls too, rather than with the `new` prefix.

#50

https://www.youtube.com/watch?v=Yz8ySbaeCf8
https://medium.com/@apalshah/javascript-class-difference-between-es5-and-es6-classes-a37b6c90c7f8
https://esdiscuss.org/topic/extending-an-es6-class-using-es5-syntax