Int8 Migration
Demoed this one out in the real implementation first, just to see if it is fully compatible with the rest of the library, in terms of typings and the runtime implementation itself.

For some reason it looks like the typings for this don't properly persist the `valueOf()` return type generic, I don't want it to default to `number` as the type, but it seems like it is.