toPrimitive Demo Fix
Possibly seems to be another possible fix for type issues with the return type for `valueOf()`. I already tried Excluding `valueOf()` from the `T` number type, but that makes it no longer a valid `number` type, hence breaking all of the other usage points that come with that in particular. That part of the branded type is crucial to why things work altogether.

https://javascript.info/object-toprimitive