Primitive Typings for Classes
I think I found an actually fairly nice way to handle this!

I can assert the exported types of the classes, and make them more accurate to how they are used, rather than specific to how they are implemented. Going to look into this now, it seems to look like it works very well! Using ES6 classes still is a big plus as well, as it still works with instanceof checks.

TypeScript doesn't like `instanceof` calls on non-object or `never` values though, so if needed I can implement a static method maybe, like arrays? It seems to be okay with using `instanceof` on an `unknown` value, which is good.

```ts
  static isInt8(value: unknown): value is Int8 {
    return value instanceof Int8;
  }
```

I also really like this custom types export setup because it doesn't change the usage of existing programs, and it's only at the type level too, so things functionally won't change at all at the runtime level.

Flagman, Lifesigns!

#50

This commit is essentially a revert back to the start of this branch, then starting new with this custom approach.