Number Extension Types
Realized that the constructor didn't have the proper typings that the class did, since it extends `Number`. Now I think I get it, that you have to extend both the instance prototype, as well as the constructor prototype, hence for static methods. So this is key to properly reflect in the types as well :)

Classes do this in TS for us out of the box, but I'm doing things particular here.

https://stackoverflow.com/questions/58562612/create-a-derived-type-from-class-but-omit-the-constructor-typescript