Simple Primitive Typings
Not sure how I feel about going back to something like this, but it does seem pretty cool. I could make the public API specific in terms of how the types are implemented, rather than having to trim the internal implementation to be small for the public API.

This is a very nice contrast against just listening to Code Orange haha.
https://hotelpoolsmusic.bandcamp.com/album/evenings-vol-1-2