Strict Branded Typings Unvert
Going back on this loose typing change, just because I think it's still important to strictly type the internals of your implementation, even if the exported parts are more generic. That's probably even moreso why the internals should be strictly-typed, just in case you make an incorrect assertion somewhere along the way.