Read Adjacent Demo
Been working on this implementation for a little while now, it is an abstraction layer that hides away the iteration detection for reading adjacently-placed root tags that are in a single file. This is something that has to be implemented for Bedrock and 3DS Edition.

Originally I was going to add this as a first-party function in NBTify itself, but now I think I like the idea of instead just exposing this underlying metadata to the read function itself, and allowing users to build the abstractions they need, specific to the files they need to parse. Since this is more specific to individual NBT formats, and it's really inherent that it works properly, I don't want to obscure the concerns for that. This does appear to work awesome actually though! I just don't want to generalize this so where it will 'work for everything', but not be easy to use. That was the initial part of NBTify's structure, to allow you to build your own primitives to work with. So `read()` for example is broad enough to work for any format, and I don't want to export a more narrow function alongside that, when you can just narrow the usage of `read()` itself in your own code. Essentially, I want it to be 'unopinionated' in how it lets you use it. Of course that in and of itself is an opinion on how to do it haha.