SNBT Parsing Re-Class
Now the SNBT parser is a class again! I really love functional programming. However, if there is state involved, and it's kind of related to do with instancing, then maybe it should be a class. If JavaScript had pointers, that would be one thing, but the only way to achieve something similar is with wrapping it in an object, and I don't want to deal with using that.

DTP - The Way Home!

SquiggleConf was so awesome!!! Absolutely glad I went, it was mindboggling to finally meet fellow webdevvers that know all of the same obscure things. Met a bunch of people, got a bunch of advice in how to move forwards, I just have to keep trying, and putting myself out there! Coming to this conference was a great step in doing that. It was also that I should focus on honing my most strong skills, rather than trying to be a little good at a whole bunch of different things. TypeScript and web dev in general do seem to be that in terms of coding for me right now.