New SNBT Parser Concept
Things just started clicking with this the other day, this was while on the plane back home from SquiggleConf. I didn't even get to run it or try it actually since it's not complete yet. This was a nice effort where I did start understanding how I could do it a bit better. I don't think it's quite how I want to move forwards with it though, so I'm just going to save it as an attempt at learning it. I found an example of writing your own JSON parser in TS, and it seems to follow paradigms I really like, and would want to modify/build similarly for my needs here.

Gonna make note of those in the new commits for that, but here they are just in case, for reference. They weren't what I used for this commit.

https://github.com/teamchong/json-literal-types
https://medium.com/@anchen.li/writing-a-json-parser-in-typescript-72c90fe6df8c
https://github.com/mzbac/JSON-parser

I think I heard about the types one first in the MiTS Discord while I happened to be going through it, and it kind of made things click for me, that a JSON parser would be very similar to that of an SNBT one, just with a few modifications :)