NBT XML?
Looking into being able to store NBT as XML! I think I'm just curious how this might be able to be interesting in terms of possibly reactive components, and or JSX, and just composability in general. You wouldn't want to write it out by hand like this, but maybe instead it could be a way to render NBT in a vector format, say like with SVG :O

#46

Got some help from GPT and the web in general (along with some of my own validations) to get what would be the XML entities for the odd characters I have experimented with serializing, like newlines, tabs, things like those.

This file was written by hand, as a transform from the original `bigtest.snbt` file. I am going to see if I can make a custom parser and writer for XML, which goes between NBTify JS primitives, and XML strings. Seems like it could be another great way to make NBT interop even greater!