XML Parsing!
Tiny Tears, Terria

https://www.npmjs.com/package/xmldom

Getting some help from GPT to implement my own XML to NBTify object parser, since it is essentially tag-specific in what it parses to. Thankfully it actually does make a bit of sense to use an existing XML tree parser, and to generate the NBT object tree from that. I wonder if that's not very performant though, maybe handling it all on my own would make more sense in terms of being scoped to NBTify. I don't like needing to parse XML outright with my own code, but I also don't want to need to rely on generating actual DOM trees themselves, unless maybe that makes sense to do so. Seems like it could be a bunch of extra memory overhead for no reason, if just converting to and from NBT. If you want to render it as a tree of elements in the browser though, it might actually make a lot of sense to do so like that though. Would it still make sense to go from NBT, to a JS object tree, then to an XML element tree? I kind of like the premise of that a bit more. We'll see though!

Kind of surprised this works so well actually, GPT got things working not too bad, and considering this is in Node as well! So DOMParser is from the polyfill itself.

And actually, now I want to make my own DOMParser code but in TypeScript, maybe I can fork xmldom and add types, and make things more concise overall.

#46