VarInt Fixes 🚀
Bringing the config for the project up.

Tried to add support for isolatedDeclarations, but it wasn't too different for things, and my tests aren't too friendly for it yet, so I'm not going to do it for this one yet.

I think I actually should have a separate tsconfig for the tests folder come to think of it, that would work good.

#53