Optional Root Check Demo
Trying out a few options for this. Either the plain read/write APIs can specifically handle if a root tag should be validated, and it can be an argument to these same original functions, or it could be a new function you import to specifically validate whether it is a root tag or not.

This is more a demo/test and placeholder to see how this might work.

I still have to adjust the options to work with writing too.

```sh
echo '[B;1b]' | nbtify --anyroot
```

#54