Set Polyfill Replacer-Reviver
#33

So this is awesome! For some reason adding a transform for Uint8Array objects wasn't working. Having a custom setup for `Set` objects appears to be working though.

This makes me think actually, this configuration would actually work awesome with things like `FlatWorldLayers` in the `level.dat` for example, because you would just add a transform in there that adds an extra call to `JSON.stringify()` and friends. So having plain JS trees makes even more sense in this case :D

The end of my song 'Scopes' sounds a lot like Clown Core come to think of it!

Also, The Land of the Dear feels like a reference back to a bunch of my older projects, which makes sense why it feels so near and dear to me (pun unintended). It has a little bit of foreshadowing to my sound later on too.