Reviver Debug
Checking that the order properties are logged is in the same order that JSON.parse() logs them. Stringifying should be BFS, while parsing should be DFS. It's the order that they are traversed as the file is read/written through.

In theory this means that you would be able to write a replacer/reviver pair and it would work both with JSON and NBT, if you did things in a certain way. This seems like a very neat thing that could make converting between the formats much more safe. It seems like a natural progression for adding custom primitives to either format, by way of essentially transpiling them in.

Come to think of it, this would actually work very well for converting to and from NBTify primitives to Prismarine-NBT objects as well, that might be a nice way to add that to NBTify itself without needing to implement custom functions or options for that outright.

#33
#22