Transform `this` Contexts
Debugging how the values should be completed when the transforms are applied. It seems like it should be a second iteration, rather than on the initial parse loop. So the `read()` reviver is currently incorrect, because as you get each value, it appends it and calls the reviver on it, but the parent doesn't have all of the children in it yet. JSON appears to have all of the children in the parent before it iterates over the value for it's individual keys.

```js
JSON.parse('{ "hi": { "nice": [1,2,3,4,5] } }', function(key, value) { console.log(this, key, value); return value; });
```

Also, do I need to add a transform for TypedArray-like tags? I feel like not, because they can only ever include their intended child types, which are numbers. In the event you could have custom values inside of a `Uint8Array` for some reason, then yes I would want to add a transform, but otherwise I'm not sure what you would transform. Hang on, actually, that enables the whole premise of being able to change TypedArray values to add custom tags. I want to do that! So you technically could embed custom file trees into TAG_Byte_Array values for example, and the transform functions would be able to detect whether it's in a specific format or not. This seems like a good way to do things :)