Simpler File Demo
The content in `bigtest.snbt` is more specific to what I want to test, the values in the regular file are a lot longer, and it doesn't have all of the array tag types. Ideally I will just pass the transforms into the SNBT methods themselves, but that isn't implemented just yet.

Lifesigns - Lifesigns