FlatWorldLayers Transform Demo/Test
This is a great example of why NBT parsing with transforms is such a cool thing (I'm just now realizing the implications of this!!). It allows you to easily make the data into the shape you need, as you parse the data itself, without needing to chain your own function calls and things like that. Having callbacks for hooking into the process itself is even more powerful than that, and I'm starting to see a bunch more programming paradigms I had never thought of before.